requirements.txt
absl-py==0.8.0
astor==0.8.0
certifi==2019.9.11
chardet==3.0.4
Django==2.2.6
future==0.17.1
gast==0.2.2
google-pasta==0.1.7
grpcio==1.24.0
h5py==2.10.0
idna==2.8
joblib==0.14.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
line-bot-sdk==1.14.0
Markdown==3.1.1
mecab-python3==0.996.2
numpy==1.17.2
opt-einsum==3.1.0
pandas==0.25.1
protobuf==3.9.2
python-dateutil==2.8.0
pytz==2019.2
requests==2.22.0
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
sqlparse==0.3.0
tensorboard==2.0.0
tensorflow==2.0.0
tensorflow-estimator==2.0.0
termcolor==1.1.0
urllib3==1.25.6
Werkzeug==0.16.0
wrapt==1.11.2
